Add lazy utils import
